organized the Button component files and change the padding
