+++ src/components/Button/Button.component.styles.tsx
@@ -1,0 +1,145 @@
+import styled, { CSSObject } from "@emotion/styled";
+import { colorPalette} from '../../theme';
+import { typography } from '../../theme';
+import { ButtonOption } from "./Button.component.types";
+
+const base: CSSObject = {
+  color:colorPalette.Text.White,
+  cursor: "pointer",
+  display: "flex",
+  fontWeight: "600",
+  alignItems: "center",
+  justifyContent: "center",
+  verticalAlign: "middle",
+  userSelect: "none",
+  border: "2px solid transparent",
+  fontSize: "20px",
+  gap:"10px",
+  transition: "all .15s ease-in-out",
+  "&:focus": {
+    outline: "0",
+  },
+};
+
+type EnumStyles = Record<string, CSSObject>;
+
+const buttonVariant: EnumStyles = {
+  primary: {
+    color:colorPalette.Text.White,
+    backgroundColor: colorPalette.DeviDaisy[500],
+    boxShadow: "5px 10px 20px rgba(0,0,0,0.25)",
+    borderRadius: "8px",
+    ":hover":{
+      boxSizing:"border-box",
+      color:colorPalette.DeviDaisy[500],
+      backgroundColor: "transparent",
+      border: "2px solid #6044B5",
+      boxShadow: "(5px 10px 20px rgba(0,0,0,0.25)",
+      borderRadius: "8px",
+    },
+  },
+  error: {
+    color: colorPalette.Text.White,
+    backgroundColor: colorPalette.Status.Error,
+    boxShadow: "5px 10px 20px rgba(0,0,0,0.25)",
+    borderRadius: "8px",
+    ":hover":{
+      boxSizing:"border-box",
+      color:colorPalette.Status.Error,
+      backgroundColor: "transparent",
+      border: "2px solid #C12945",
+      boxShadow: "(5px 10px 20px rgba(0,0,0,0.25)",
+      borderRadius: "8px",
+    },
+  },
+  positive: {
+    color: colorPalette.Text.White,
+    backgroundColor: colorPalette.Status.Positive,
+    boxShadow: "5px 10px 20px rgba(0,0,0,0.25)",
+    borderRadius: "8px",
+    ":hover":{
+      boxSizing:"border-box",
+      color:colorPalette.Status.Positive,
+      backgroundColor: "transparent",
+      border: "2px solid #26962B",
+      boxShadow: "(5px 10px 20px rgba(0,0,0,0.25)",
+      borderRadius: "8px",
+    },
+  },
+};
+
+const buttonSize: EnumStyles = {
+  small: {
+    margin:"auto",
+    width:"103px",
+    height:"35px",
+    fontFamily: typography.Quicksand,
+    fontWeight:'500',
+    fontSize: "14px",
+    padding: "5px 30px"
+  },
+  medium: {
+    margin:"auto",
+    width:"150px",
+    height:"50px",
+    fontFamily: typography.Quicksand,
+    fontWeight:'600',
+    fontSize: "20px",
+    padding: "15px 40px",
+  },
+  large: {
+    margin:"auto",
+    width:"257px",
+    height:"73px",
+    fontFamily: typography.Quicksand,
+    fontWeight:'600',
+    fontSize: "20px",
+    padding: "20px 80px",
+  },
+};
+
+const buttonDisabled: EnumStyles = {
+   primary: {
+    cursor:"not-allowed",
+    color:colorPalette.Text.White,
+    backgroundColor: colorPalette.DeviDaisy[500],
+    background: "rgba(96, 68, 181, 0.35)",
+    boxShadow: "5px 10px 20px rgba(0, 0, 0, 0.25)",
+    borderRadius: "8px",
+  },
+  error: {
+    cursor:"not-allowed",
+    color:colorPalette.Text.White,
+    backgroundColor: colorPalette.Status.Error,
+    background: "rgba(193, 41, 69, 0.55)",
+    boxShadow: "5px 10px 20px rgba(0, 0, 0, 0.25)",
+    borderRadius: "8px",
+  }, 
+  positive: {
+    cursor:"not-allowed",
+    color:colorPalette.Text.White,
+    backgroundColor: colorPalette.Status.Positive,
+    background: "rgba(38, 150, 43, 0.5)",
+    boxShadow: "5px 10px 20px rgba(0, 0, 0, 0.25)",
+    borderRadius: "8px",
+  },
+};
+
+
+
+const fullWidth = ({ isFullWidth = false }: ButtonOption) =>
+  isFullWidth
+    ? {
+        display: "block",
+        width: "100%",
+      }
+    : null;
+
+  
+export const ButtonComponent = styled.button<ButtonOption>`
+      ${base},
+      ${(props) => props.size && buttonSize[props.size]}
+      ${(props) => props.disabled ? buttonDisabled[props.variant]  : buttonVariant[props.variant]}
+      ${fullWidth}
+  `;
+
