+++ src/components/Button/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Button";
+export { default } from "./Button.component";
